Fix: global/shared context initialization/deinitialization
