+++ src/Interpreters/Context.h
@@ -25,6 +25,7 @@
 #include <Server/HTTP/HTTPContext.h>
 #include <Storages/ColumnsDescription.h>
 #include <Storages/IStorage_fwd.h>
+#include <Interpreters/ContextShared_fwd.h>
 
 #include "config.h"
 
@@ -48,7 +49,6 @@ namespace DB
 
 class ASTSelectQuery;
 
-struct ContextSharedPart;
 class ContextAccess;
 class ContextAccessWrapper;
 struct User;
@@ -232,24 +232,6 @@ struct IHostContext
 };
 
 using IHostContextPtr = std::shared_ptr<IHostContext>;
-
-/// A small class which owns ContextShared.
-/// We don't use something like unique_ptr directly to allow ContextShared type to be incomplete.
-struct SharedContextHolder
-{
-    ~SharedContextHolder();
-    SharedContextHolder();
-    explicit SharedContextHolder(std::unique_ptr<ContextSharedPart> shared_context);
-    SharedContextHolder(SharedContextHolder &&) noexcept;
-
-    SharedContextHolder & operator=(SharedContextHolder &&) noexcept;
-
-    ContextSharedPart * get() const { return shared.get(); }
-    void reset();
-
-private:
-    std::unique_ptr<ContextSharedPart> shared;
-};
 
 class ContextSharedMutex : public SharedMutexHelper<ContextSharedMutex>
 {
