+++ src/Interpreters/ContextShared_fwd.h
@@ -1,0 +1,27 @@
+#pragma once
+
+#include <memory>
+
+namespace DB
+{
+struct ContextSharedPart;
+
+/// A small class which owns ContextShared.
+/// We don't use something like unique_ptr directly to allow ContextShared type to be incomplete.
+struct SharedContextHolder
+{
+    ~SharedContextHolder();
+    SharedContextHolder();
+    explicit SharedContextHolder(std::unique_ptr<ContextSharedPart> shared_context);
+    SharedContextHolder(SharedContextHolder &&) noexcept;
+
+    SharedContextHolder & operator=(SharedContextHolder &&) noexcept;
+
+    ContextSharedPart * get() const { return shared.get(); }
+    void reset();
+
+private:
+    std::unique_ptr<ContextSharedPart> shared;
+};
+
+}
